# frontend/src/lib/backend/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "Sur la route", "version": "0.1.0" },
	"paths": {
...
		"/api/users/": {
			"get": {
				"tags": ["users"],
				"summary": "Read",
				"description": "Read users.",
				"operationId": "users-read",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "skip",
						"in": "query",
						"required": false,
						"schema": { "type": "integer", "default": 0, "title": "Skip" }
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": { "type": "integer", "default": 100, "title": "Limit" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UsersPublic" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"post": {
				"tags": ["users"],
				"summary": "Create",
				"description": "Create new user.",
				"operationId": "users-create",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } }
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/users/me": {
			"get": {
				"tags": ["users"],
				"summary": "Read Me",
				"description": "Get current user.",
				"operationId": "users-read_me",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
						}
					}
				},
				"security": [{ "OAuth2PasswordBearer": [] }]
			},
			"delete": {
				"tags": ["users"],
				"summary": "Delete Me",
				"description": "Delete own user.",
				"operationId": "users-delete_me",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
						}
					}
				},
				"security": [{ "OAuth2PasswordBearer": [] }]
			},
			"patch": {
				"tags": ["users"],
				"summary": "Update Me",
				"description": "Update own user.",
				"operationId": "users-update_me",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/UserUpdateMe" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				},
				"security": [{ "OAuth2PasswordBearer": [] }]
			}
		},
		"/api/users/me/password": {
			"patch": {
				"tags": ["users"],
				"summary": "Update Password Me",
				"description": "Update own password.",
				"operationId": "users-update_password_me",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/UpdatePassword" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				},
				"security": [{ "OAuth2PasswordBearer": [] }]
			}
		},
		"/api/users/signup": {
			"post": {
				"tags": ["users"],
				"summary": "Register",
				"description": "Create new user without the need to be logged in.",
				"operationId": "users-register",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/UserRegister" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/api/users/{user_id}": {
			"get": {
				"tags": ["users"],
				"summary": "Read By Id",
				"description": "Get a specific user by id.",
				"operationId": "users-read_by_id",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "User Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"patch": {
				"tags": ["users"],
				"summary": "Update",
				"description": "Update a user.",
				"operationId": "users-update",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "User Id" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" } }
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserPublic" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["users"],
				"summary": "Delete",
				"description": "Delete a user.",
				"operationId": "users-delete",
				"security": [{ "OAuth2PasswordBearer": [] }],
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid", "title": "User Id" }
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/Message" } }
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"Message": {
				"properties": { "message": { "type": "string", "title": "Message" } },
				"type": "object",
				"required": ["message"],
				"title": "Message"
			},
...
			"UpdatePassword": {
				"properties": {
					"current_password": {
						"type": "string",
						"maxLength": 40,
						"minLength": 8,
						"title": "Current Password"
					},
					"new_password": {
						"type": "string",
						"maxLength": 40,
						"minLength": 8,
						"title": "New Password"
					}
				},
				"type": "object",
				"required": ["current_password", "new_password"],
				"title": "UpdatePassword"
			},
...
			"UserCreate": {
				"properties": {
					"email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
					"is_active": { "type": "boolean", "title": "Is Active", "default": true },
					"is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
					"full_name": {
						"anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
						"title": "Full Name"
					},
					"password": { "type": "string", "maxLength": 40, "minLength": 8, "title": "Password" }
				},
				"type": "object",
				"required": ["email", "password"],
				"title": "UserCreate"
			},
			"UserPublic": {
				"properties": {
					"email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
					"is_active": { "type": "boolean", "title": "Is Active", "default": true },
					"is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
					"full_name": {
						"anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
						"title": "Full Name"
					},
					"id": { "type": "string", "format": "uuid", "title": "Id" }
				},
				"type": "object",
				"required": ["email", "id"],
				"title": "UserPublic"
			},
			"UserRegister": {
				"properties": {
					"email": { "type": "string", "maxLength": 255, "format": "email", "title": "Email" },
					"password": { "type": "string", "maxLength": 40, "minLength": 8, "title": "Password" },
					"full_name": {
						"anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
						"title": "Full Name"
					}
				},
				"type": "object",
				"required": ["email", "password"],
				"title": "UserRegister"
			},
			"UserUpdate": {
				"properties": {
					"email": {
						"anyOf": [
							{ "type": "string", "maxLength": 255, "format": "email" },
							{ "type": "null" }
						],
						"title": "Email"
					},
					"is_active": { "type": "boolean", "title": "Is Active", "default": true },
					"is_superuser": { "type": "boolean", "title": "Is Superuser", "default": false },
					"full_name": {
						"anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
						"title": "Full Name"
					},
					"password": {
						"anyOf": [{ "type": "string", "maxLength": 40, "minLength": 8 }, { "type": "null" }],
						"title": "Password"
					}
				},
				"type": "object",
				"title": "UserUpdate"
			},
			"UserUpdateMe": {
				"properties": {
					"full_name": {
						"anyOf": [{ "type": "string", "maxLength": 255 }, { "type": "null" }],
						"title": "Full Name"
					},
					"email": {
						"anyOf": [
							{ "type": "string", "maxLength": 255, "format": "email" },
							{ "type": "null" }
						],
						"title": "Email"
					}
				},
				"type": "object",
				"title": "UserUpdateMe"
			},
			"UsersPublic": {
				"properties": {
					"data": {
						"items": { "$ref": "#/components/schemas/UserPublic" },
						"type": "array",
						"title": "Data"
					},
					"count": { "type": "integer", "title": "Count" }
				},
				"type": "object",
				"required": ["data", "count"],
				"title": "UsersPublic"
			},
			"ValidationError": {
				"properties": {
					"loc": {
						"items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
						"type": "array",
						"title": "Location"
					},
					"msg": { "type": "string", "title": "Message" },
					"type": { "type": "string", "title": "Error Type" }
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			}
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": { "password": { "scopes": {}, "tokenUrl": "/api/login/access-token" } }
			}
		}
	}
}